Move 'v2' token_storage up a level
All of the 'v2' entities move up one nesting level, now that the 'v1'
items are renamed as 'legacy'.
