+++ src/globus_sdk/token_storage/memory.py
