+++ tests/unit/globus_app/test_authorizer_factory.py
@@ -13,7 +13,7 @@ from globus_sdk.token_storage import (
     MemoryTokenStorage,
     NotExpiredValidator,
 )
-from globus_sdk.token_storage.v2.validating_token_storage import (
+from globus_sdk.token_storage.validating_token_storage import (
     ExpiredTokenError,
     MissingTokenError,
     ValidatingTokenStorage,
