+++ tests/unit/tokenstorage/v2/test_token_storage.py
@@ -1,7 +1,7 @@
 import pytest
 
 from globus_sdk import GlobusSDKUsageError
-from globus_sdk.token_storage.v2.base import _slugify_app_name
+from globus_sdk.token_storage.base import _slugify_app_name
 
 
 @pytest.mark.parametrize(
